+++ ReleaseNotes/release_notes.md
@@ -1,11 +1,16 @@
 ## New features:
   1. add `Ostap::Kinematics::Dalitz0::E1,E2,E3` and `Ostap::Kinematics::Dalitz0::P1,P2,P3` three-argument methods, They are less efficient than corresponding two-argument methods of class `Ostap:Kinematics::Dalitz`
+  1. Fix `Ostap::Math::PSDalitz`
+  1. Upgrade `Ostap::Math::PhaseSpace23L`
 
 
 ## Backward incompatible  
 
 ## Bug fixes:
 
+   1. Fix `Ostap::Kinematics::phasepace3` for cases with some arguments are zero
+ 
+ 
 # v1.6.6.0
 
 ## New features: